+++ test/test_doc.py
@@ -1,5 +1,6 @@
 import builtins
 import dataclasses
+import sys
 from pathlib import Path
 from unittest.mock import patch
 
@@ -86,3 +87,16 @@ def test_class_with_raising_getattr():
 def test_builtin_source_file():
     m = Module(builtins)
     assert m.source_file is None
+
+
+@pytest.mark.skipif(sys.version_info < (3, 9), reason="3.9+ only")
+def test_raising_getdoc():
+    class Foo:
+        @classmethod
+        @property
+        def __doc__(self):
+            raise RuntimeError
+
+    x = Class(Foo.__module__, Foo.__qualname__, Foo, (Foo.__module__, Foo.__qualname__))
+    with pytest.warns(RuntimeWarning, match="inspect.getdoc(.+) raised an exception"):
+        assert x.docstring == ""